tech(api): move scoring-config-repo and models related to scoring from shared to evaluation bounded context
