+++ api/src/certification/evaluation/infrastructure/repositories/scoring-configuration-repository.js
