+++ api/tests/tooling/domain-builder/factory/certification/shared/build-v3-certification-scoring.js
@@ -1,4 +1,4 @@
-import { V3CertificationScoring } from '../../../../../../src/certification/shared/domain/models/V3CertificationScoring.js';
+import { V3CertificationScoring } from '../../../../../../src/certification/evaluation/domain/models/V3CertificationScoring.js';
 
 export const buildV3CertificationScoring = ({
   competencesForScoring = [],
